add: convert main functionality from pages router to app router
